up/requirements.txt
asgiref==3.4.1
boto3==1.18.47
botocore==1.21.47
dj-database-url==0.5.0
Django==3.2.7
django-environ==0.7.0
django-filter==21.1
django-storages==1.11.1
djangorestframework==3.12.4
gunicorn==20.1.0
jmespath==0.10.0
Markdown==3.3.4
mysqlclient==2.0.3
Pillow==8.3.2
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
